refactor: move IT hygiene constants to a dedicated file for better organization
