+++ plugins/wazuh-analysis/public/groups/security-operations/apps/it-hygiene/application.tsx
@@ -4,8 +4,9 @@ import ReactDOM from 'react-dom';
 import { SecurityOperationsNavGroup } from '../..';
 import { Layout } from '../../../layout';
 import { createSideNavItems } from '../../../side-nav';
-import { IT_HYGIENE_ID, SECURITY_OPERATIONS_TITLE } from '../../constants';
+import { SECURITY_OPERATIONS_TITLE } from '../../constants';
 import { ItHygieneApp } from './it-hygiene-app';
+import { IT_HYGIENE_ID } from './constants';
 
 export const renderApp = async (params: AppMountParameters) => {
   const items = createSideNavItems({
