+++ plugins/wazuh-analysis/public/groups/security-operations/apps/it-hygiene/it-hygiene-app.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AppMountParameters } from 'opensearch-dashboards/public';
-import { IT_HYGIENE_TITLE } from '../../constants';
+import { IT_HYGIENE_TITLE } from './constants';
 
 interface ItHygieneAppProps {
   params: AppMountParameters;
