+++ plugins/wazuh-analysis/public/groups/security-operations/constants.ts
@@ -20,19 +20,9 @@ export const REGULATORY_COMPLIANCE_ID = buildSubAppId(
   SECURITY_OPERATIONS_ID,
   'regulatory_compliance',
 );
-export const IT_HYGIENE_ID = buildSubAppId(
-  SECURITY_OPERATIONS_ID,
-  'it_hygiene',
-);
 export const REGULATORY_COMPLIANCE_TITLE = i18n.translate(
   `${PLUGIN_ID}.category.${REGULATORY_COMPLIANCE_ID}`,
   {
     defaultMessage: 'Regulatory Compliance',
-  },
-);
-export const IT_HYGIENE_TITLE = i18n.translate(
-  `${PLUGIN_ID}.category.${IT_HYGIENE_ID}`,
-  {
-    defaultMessage: 'IT Hygiene',
   },
 );
